AN-310513 Fix Dragging from Design Library in Interface Designer
Dragging an Interface from the Design Library would throw an exception
trying to get the templateComponentType of the Interface. While
out-of-the-box components have this field set, Interface in the Design
Library do not. Updating the indexing to use get allows it to default to
None when dragging from the Design Library and work correctly.